+++ src/audio_decoder/ad_plugin.h
@@ -8,14 +8,14 @@ typedef struct {
 	int     (*close)(void *);
 	int     (*info)(void *, struct adinfo *);
 	int64_t (*seek)(void *, int64_t);
-	ssize_t (*read_dbl)(void *, double *, size_t);
+	ssize_t (*read_dbl)(void *, float *, size_t);
 } ad_plugin;
 
 void *  ad_open_null(const char *, struct adinfo *);
 int     ad_close_null(void *);
 int     ad_info_null(void *, struct adinfo *);
 int64_t ad_seek_null(void *, int64_t);
-ssize_t ad_read_null(void *, double*, size_t);
+ssize_t ad_read_null(void *, float*, size_t);
 
 /* hardcoded backends */
 const ad_plugin * get_sndfile();
